Fix background scrape on enter
It will now read the correct session value.

Also moves some session helpers to a separate file.
